Refactor: move render passes to Renderer.Passes namespace
Reorganized all render pass classes (e.g., BloomPass, CSMPass, GBufferPass, etc.) into the new SponzaScene.Renderer.Passes namespace. Updated file headers and using directives accordingly. No functional changes; this improves code structure and maintainability.
