+++ sources/Experiments/SponzaScene/Renderer/Passes/ComposePass.cs
@@ -2,7 +2,7 @@
 using Zenith.NET;
 using Buffer = Zenith.NET.Buffer;
 
-namespace SponzaScene.Renderer;
+namespace SponzaScene.Renderer.Passes;
 
 internal unsafe class ComposePass : FullscreenPass
 {
