+++ sources/Experiments/SponzaScene/Renderer/Passes/VolumetricLightBlurPass.cs
@@ -5,7 +5,7 @@ using Zenith.NET;
 using Zenith.NET.Extensions.Slang;
 using Buffer = Zenith.NET.Buffer;
 
-namespace SponzaScene.Renderer;
+namespace SponzaScene.Renderer.Passes;
 
 internal unsafe class VolumetricLightBlurPass : RenderPass
 {
